Use Jython directly in validateModel for faster performance (#506)
* Initial check-in, read alias files, find folders for paths

* Set system property to avoid warning in JDK 11

* Remove wlst.silence(), not needed for Jython

* Move non-WLST methods to model helper

* Move non-WLST methods to model helper

* Revised scripts to use Jython directly

* Remove wlst.silence(), not needed in Jython

* Remove Java UsagePrinter

* Set python console to empty for 14.1.1
