+++ core/src/main/python/wlsdeploy/util/cla_helper.py
@@ -1,5 +1,5 @@
 """
-Copyright (c) 2019, Oracle Corporation and/or its affiliates.  All rights reserved.
+Copyright (c) 2019, 2020, Oracle Corporation and/or its affiliates.  All rights reserved.
 Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
 Utility CLS methods shared by multiple tools.
@@ -9,9 +9,9 @@ from oracle.weblogic.deploy.util import FileUtils
 
 from wlsdeploy.exception import exception_helper
 from wlsdeploy.logging.platform_logger import PlatformLogger
-from wlsdeploy.tool.deploy import deployer_utils
 from wlsdeploy.tool.util.archive_helper import ArchiveHelper
 from wlsdeploy.util import cla_utils
+from wlsdeploy.util import model_helper
 from wlsdeploy.util import variables
 from wlsdeploy.util.cla_utils import CommandLineArgUtil
 from wlsdeploy.util.model_translator import FileToPython
@@ -166,7 +166,7 @@ def _merge_dictionaries(dictionary, new_dictionary, variable_map):
         # the new key should replace the existing one - delete the existing key and add the new one
         elif replace_key:
             del dictionary[dictionary_key]
-            if not deployer_utils.is_delete_name(new_key):
+            if not model_helper.is_delete_name(new_key):
                 dictionary[new_key] = new_value
 
         # the key is in both dictionaries - merge if the values are dictionaries, otherwise replace the value
@@ -192,11 +192,11 @@ def _find_dictionary_merge_key(dictionary, new_key, variable_map):
     if new_key in dictionary:
         return new_key, False
 
-    new_is_delete = deployer_utils.is_delete_name(new_key)
+    new_is_delete = model_helper.is_delete_name(new_key)
     match_new_key = _get_merge_match_key(new_key, variable_map)
 
     for dictionary_key in dictionary.keys():
-        dictionary_is_delete = deployer_utils.is_delete_name(dictionary_key)
+        dictionary_is_delete = model_helper.is_delete_name(dictionary_key)
         match_dictionary_key = _get_merge_match_key(dictionary_key, variable_map)
         if match_dictionary_key == match_new_key:
             replace_key = new_is_delete != dictionary_is_delete
@@ -218,6 +218,6 @@ def _get_merge_match_key(key, variable_map):
     else:
         match_key = key
 
-    if deployer_utils.is_delete_name(match_key):
-        match_key = deployer_utils.get_delete_item_name(match_key)
+    if model_helper.is_delete_name(match_key):
+        match_key = model_helper.get_delete_item_name(match_key)
     return match_key
